+++ packages/editor-core/src/nodeviews/index.ts
@@ -4,6 +4,6 @@ export { default as ReactMediaGroupNode } from './ui/media-group';
 export { default as ReactMediaNode } from './ui/media';
 export { default as ReactMentionNode } from './ui/mention';
 export { panelNodeView } from './ui/panel';
+export { decisionItemNodeView } from './ui/decisionItem';
 
 export { ProsemirrorGetPosHandler } from './ui';
-export { default as codeMirrorNodeView } from './ui/code-mirror';