+++ packages/editor-core/src/plugins/index.ts
@@ -1,7 +1,6 @@
 export { default as blockTypePlugins, stateKey as blockTypeStateKey } from './block-type';
 export { default as clearFormattingPlugins, stateKey as clearFormattingStateKey } from './clear-formatting';
 export { default as codeBlockPlugins, stateKey as codeBlockStateKey, CodeBlockState } from './code-block';
-export { default as codeMirrorPlugins } from './code-mirror';
 export { default as emojisPlugins, stateKey as emojisStateKey } from './emojis';
 export { default as asciiEmojiPlugins, stateKey as asciiEmojiStateKey } from './emojis/ascii-input-rules';
 export { default as hyperlinkPlugins, stateKey as hyperlinkStateKey } from './hyperlink';
@@ -15,3 +14,4 @@ export { default as rulePlugins, stateKey as ruleStateKey } from './rule';
 export { default as panelPlugins, stateKey as panelStateKey } from './panel';
 export { default as reactNodeViewPlugins, ReactNodeViewState, stateKey as reactNodeViewStateKey } from './react-nodeview';
 export { default as tablePlugins, TableState, stateKey as tableStateKey } from './table';
+export { default as tasksAndDecisionsPlugin } from './tasks-and-decisions';