+++ packages/editor-core/src/plugins/media/keymap.ts
@@ -9,6 +9,8 @@ export function keymapPlugin(schema: Schema<any, any>): Plugin {
   const list = {};
 
   keymaps.bindKeymapWithCommand(keymaps.undo.common!, ignoreLinksInSteps, list);
+  keymaps.bindKeymapWithCommand(keymaps.enter.common!, splitMediaGroup, list);
+  keymaps.bindKeymapWithCommand(keymaps.insertNewLine.common!, splitMediaGroup, list);
 
   return keymap(list);
 }
@@ -19,5 +21,9 @@ function ignoreLinksInSteps(state: EditorState<any>, dispatch: (tr: Transaction)
   return false;
 }
 
-export default keymapPlugin;
+function splitMediaGroup(state: EditorState<any>, dispatch: (tr: Transaction) => void): boolean {
+  const mediaPluginState = stateKey.getState(state) as MediaPluginState;
+  return mediaPluginState.splitMediaGroup();
+}
 
+export default keymapPlugin;