+++ packages/editor-core/stories/editor/index.tsx
@@ -8,7 +8,6 @@ import { Chrome } from '../../';
 import blockTypePlugins, { stateKey as blockTypeStateKey } from '../../src/plugins/block-type';
 import clearFormattingPlugins, { stateKey as clearFormattingStateKey } from '../../src/plugins/clear-formatting';
 import codeBlockPlugins, { stateKey as codeBlockStateKey } from '../../src/plugins/code-block';
-import codeMirrorPlugins from '../../src/plugins/code-mirror';
 import panelPlugins, { stateKey as panelStateKey } from '../../src/plugins/panel';
 import textFormattingPlugins, { stateKey as textFormattingStateKey } from '../../src/plugins/text-formatting';
 import hyperlinkPlugins, { stateKey as hyperlinkStateKey } from '../../src/plugins/hyperlink';
@@ -18,7 +17,7 @@ import mentionsPlugins, { stateKey as mentionsStateKey } from '../../src/plugins
 import emojiPlugins, { stateKey as emojiStateKey } from '../../src/plugins/emojis';
 import asciiEmojiPlugins from '../../src/plugins/emojis/ascii-input-rules';
 import tablePlugins, { stateKey as tableStateKey } from '../../src/plugins/table';
-import { reactNodeViewPlugins } from '../../src/plugins';
+import { reactNodeViewPlugins, tasksAndDecisionsPlugin } from '../../src/plugins';
 
 import textColorPlugins, { stateKey as textColorStateKey } from '../../src/plugins/text-color';
 import {
@@ -297,10 +296,10 @@ export default class Editor extends PureComponent<Props, State> {
           // this is needed until we implement keymap proirity :(
           ...listsPlugins(schema),
           ...codeBlockPlugins(schema),
-          ...codeMirrorPlugins(schema),
           ...panelPlugins(schema),
           ...tablePlugins(),
           ...reactNodeViewPlugins(schema),
+          ...tasksAndDecisionsPlugin(schema),
           history(),
           keymap(baseKeymap) // should be last :(
         ]